Implement Android Part
